requirements.txt
SQLAlchemy~=2.0.23
uvicorn~=0.24.0
numpy~=1.26.1
opencv-python~=4.8.1.78
fastapi~=0.104.1
pydantic~=2.4.2
face-recognition
python-dotenv~=1.0.0
python-jose[cryptography]